+++ components/device-types/androidsense-plugin/org.wso2.carbon.device.mgt.iot.androidsense.agent/app/src/main/java/org/wso2/carbon/iot/android/sense/event/streams/sensor/SensorDataReader.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and limitations under the License.
  *
  */
-package org.wso2.carbon.iot.android.sense.event.streams.Sensor;
+package org.wso2.carbon.iot.android.sense.event.streams.sensor;
 
 import android.content.Context;
 import android.content.SharedPreferences;
